Added Authrorization

+++ src/api/notes/handler.js
@@ -1,4 +1,4 @@
-const ClientError = require('../../exceptions/ClientError')
+const ClientError = require('../../exceptions/ClientError');
 
 class NotesHandler {
   constructor(service, validator) {
@@ -15,46 +15,48 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
-      const { title = "untitled", body, tags } = request.payload;
+      const { title = 'untitled', body, tags } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
-      const noteId = await this._service.addNote({ title, body, tags });
+      const noteId = await this._service.addNote({
+        title, body, tags, owner: credentialId,
+      });
 
       const response = h.response({
-        status: "success",
-        message: "Catatan berhasil ditambahkan",
+        status: 'success',
+        message: 'Catatan berhasil ditambahkan',
         data: {
           noteId,
         },
       });
       response.code(201);
       return response;
-      } catch (error) {
-        if (error instanceof ClientError){
-          const response = h.response({
-            status: "fail",
-            message: error.message,
-          });
-          response.code(error.statusCode);
-          return response;
-        }
-        //server error
+    } catch (error) {
+      if (error instanceof ClientError) {
         const response = h.response({
-          status: "error",
-          message: "Maaf, terjadi kegagalan pada server kami.",
+          status: 'fail',
+          message: error.message,
         });
-
-        response.code(500);
-        console.error(error)
+        response.code(error.statusCode);
         return response;
       }
+
+      // Server ERROR!
+      const response = h.response({
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
+      });
+      response.code(500);
+      console.error(error);
+      return response;
+    }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
-
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
-      status: "success",
-      message: "Catatan berhasil diambil",
+      status: 'success',
       data: {
         notes,
       },
@@ -64,11 +66,12 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
 
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
-
       return {
-        status: "success",
+        status: 'success',
         data: {
           note,
         },
@@ -82,7 +85,7 @@ class NotesHandler {
         response.code(error.statusCode);
         return response;
       }
- 
+
       // Server ERROR!
       const response = h.response({
         status: 'error',
@@ -97,14 +100,14 @@ class NotesHandler {
   async putNoteByIdHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
+      const { id } = request.params; const { id: credentialId } = request.auth.credentials;
 
-      const { id } = request.params;
-
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.editNoteById(id, request.payload);
 
       return {
-        status: "success",
-        message: "Catatan berhasil diperbarui",
+        status: 'success',
+        message: 'Catatan berhasil diperbarui',
       };
     } catch (error) {
       if (error instanceof ClientError) {
@@ -115,7 +118,7 @@ class NotesHandler {
         response.code(error.statusCode);
         return response;
       }
- 
+
       // Server ERROR!
       const response = h.response({
         status: 'error',
@@ -130,12 +133,14 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
 
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.deleteNoteById(id);
 
       return {
-        status: "success",
-        message: "Catatan berhasil dihapus",
+        status: 'success',
+        message: 'Catatan berhasil dihapus',
       };
     } catch (error) {
       if (error instanceof ClientError) {
@@ -146,7 +151,7 @@ class NotesHandler {
         response.code(error.statusCode);
         return response;
       }
- 
+
       // Server ERROR!
       const response = h.response({
         status: 'error',
